Add active community to Criteria
For a project to move from Incubation to Full Stewardship, it must have
more than the code quality criteria currently enumerated, it must also
have a high enough level of dimensions such as community and process.

This commit is a first sketch of what it means to have an active community.