Minor phrasing changes in introduction and criteria pages
Rewords reference in introduction to remove ambiguity.
"Which" could refer to both the public code as the traditional
model, which can be confusing.

Rewords the example for general public in the glossary. The
previous wording was a bit confusing to me. I had to reread
it a few times before I realized that it aligned with my intuition.

Clarifies that code should be accessible without authorization.
This may be a bit pedantic, but one could imagine codebases which
are secured with a password. That would not require authentication
but would require authorization. Neither seems desirable.

Fixes a minor grammatical issue in the plain English criteria page.

Avoids repetition of document base within a sentence.