Use CI: devs do: plugable for tests
To test corner-cases like running out of disk space, or malloc
failure, structure the code to be modular enough to allow mock
failures during testing.

Blocks of code written to handle exceptional cases are some of
the least likely blocks of code to be tested ... which is very
obvious when running code coverages analysis on many codebases.