Added section "Further reading" to Introduction
It would be nice to include a few introduction links about the underlying issue of public code for further reading for people new to the subject / issue. I've added one link which is a nice whitepaper from the Free Software Foundation Europe which describe the issue of public code in great detail including interviews and case studies of public organisations working with open source/ free software explaining the value and benefits of public code.