Changed readers guide to clarify cirteria-section relation
When reading the standard for the first time it might yet not be clear how the information is structured. (my first read I thought "what criteria?)  By slightly rewriting this we can introduce the reader to the taxonomy of the standard (standard has ->criteria->section), each criteria has consistent sections which will briefly explain in this introduction.