Add new terms and footnote section to glossary.md
I've added 3 terms which are used in the standard but where not on the glossary:

Version Control / Repository / Continuous Integration. Please feel free to rewrite as in software engineering the definition of these terms is much debated (like continuous integration for example)

I also added a footnote section to be able to reference to the original source material of the definition. Git Markdown does not support footnotes so I used the HTML anchors however I can imagine this way of handling footnotes is far from ideal (not "pure" markdown, and numbering issues) I will create a separate issue for handling footnotes to refer to original source material when re-using text or definitions from other material.