Clarify versions may be derived from VCS metadata
In continuous deployment/rolling release scenario's, especially
for 'leaf' projects (less so for libraries) it is still important
to have a clear identifier for versions. However, rather than
manually tagging each version, the version would be derived from
the version control system metadata. It might be good to clarify
that explicitly.

References #419