Extract markdown as separate GH Action step
This moves the "bundle exec mdl" into a stand-alone script.
The stand-alone script is now refrenced as a separate "run"
command in both the regular test and cron-nightly actions.
As .travis.yml still references cibuild.sh, it is called
form that script as well.
Developers may still find cibuild.sh handy as a "one stop"
for testing even after we remove .travis.yml from the repo.