Add prev/next criteria links (#217, #512)
See issues #217, #512

This adds previous and next links to each criteria page, but not the
index.

In _config.yml, the syntax which was previously used to ingore the index
page was incorrect, as it would match any page starting with 'i', 'n',
'd', 'e', or 'x'. Our attempts to craft the correct way to match all but
index in this regular expression dialect failed. For now, instead, we
match all pages except those which start with an 'i', recognizing that
this may create a problem in the future.

Co-authored-by: Jan Ainali <ainali.jan@gmail.com>