MUST be in English, SHOULD be plain English
Upon re-reading it became evident that the SHOULD requirements
together comprise the concept of "plain English" for codebases
and the glossary definition needed to be exapanded or removed.

It did not make sense to put a MUST on "plain English" and a
SHOULD for all of the elements of "plain English". We believe
the spirit should clearly be a MUST for English and SHOULD be
as plain as possible.

Co-authored-by: Jan Ainali <ainali.jan@gmail.com>