Issue #557, Align cover with inner text
Prior to this commit, the cover contained the text:

> What public cod is and how to implement it for:
> P Policy makers
> M Management
> D Designers and developers

Yet, in each criterion we have a section titled:

> D Developers and designers: what you need to do

This commit makes the cover match the section headings.

Additionally, prior to this commit, when referring to
contributors, in all but two places the text had
"developers and designers"; there were two places which
listed them in the other order, which might suggest to a
reader that special emphasis for designers was needed.