Change "and" to "or" for translations
The "and" could be interpreted to mean that if you have
translations for documentation but not the code,
you are not standard compliant. And I would think that
you would almost never translate the code by changing
the name of functions and variables to non-English
even if you have a multilingual interface and
translations of the docs.