Always set PAGES_REPO_NWO in jekyll scripts
If PAGES_REPO_NWO is not set, jekyll defaults to "origin" if a remote
of that name exists, which causes link-check tests to fail with 404 for
a contributor who is not hosting a true fork. Thus if it is not set
on purpose by the caller, assume the intent is to contribute upstream
to the publiccodenet/standard repository.