Remove redundant requirement
fixes: #645

We discussed splitting the requirement:

> * All policy and source code that the codebase is based on MUST be
>   documented, reusable and portable.

in to perhaps as many as six separate requirements, as it has two
subjects (policy and source code) and three states (documented,
reusable, and portable). However we concluded that for source code these
are all clearly specified as criteria in their own right. For policy,
documented is clearly covered by another criterion, and "reusable and
portable" is far less meaningful. As such, we do not see how this adds
more value than confusion.

Co-authored-by: Jan Ainali <jan@publiccode.net>