Re-unify the CHANGELOG.md
The CHANGELOG.md was split in ffc960a3771fbc56cf2ba31cede7ad9220bb264d
in order to create the release.body text as part of the release-on-tag
workflow.

This change re-unifies the CHANGELOG.md.

The workflow now uses script/release-body.sh which parses out the
content from CHANGELOG.md or exists with error.