Improve security reporting requirements
Since https://github.com/publiccodenet/standard/issues/339 was fixed
by d0dbb84d94b3bd26427f5d2f37a0ced9687c039c the criterion has had
two very similar statements.

Previously, one statement mandated a way of reporting and the other
statement related to having documentation about how to report.

After the #339 fix, the method of reporting is no longer specified,
but the documentation of the process is a MUST, while leaving the
second requirement to document a SHOULD.

Instead we should have kept the focus of the first requirement on
the existence of a closed channel, and changed the second to a
MUST.

fixes #632

Co-authored-by: Jan Ainali <jan@publiccode.net>