Contributors SHOULD to codebase: code-in-the-open
As raised in https://github.com/publiccodenet/standard/issues/846 it is better
to write requirements that maintainers can uphold rather than focus on behavior
of contributors.

In this case, the idea is to keep sensitive information out of the codebase so
that it may be published in the open. This changes the requirement from being
about "Contributors SHOULD" to the codebase contents.