Fix the line breaks in the links
By using before instead of after for the CSS pseudomarker for the links
we get them to start on the line where the link text starts rather than
on the one it ends. It is a bit crowded in the Bundle policy and source
code criterion where we have three links in the same bullet point, but
I still think it is an overall improvement.

Fixes #306

Co-authored-by: Eric Herman  <eric@publiccode.net>