Change introduction to foreword.
This discussion has been ongoing for a long time.
I've spoken to editors and users that feel strongly about this.
I think this is a nice and minimal change that removes a few issues.

With this change the whole project becomes more adaptable.
This also creates space for a more introductory introduction later.