Link OMG specs, specific links to Further reading
fixes #944

In the future, when we have concrete examples of codebases with
policy documented in unambiguous and machine-readable formats, we
can return to the "such as this or that", but as BPMN, DMN, and CMMN
are all Object Management Group specifications, thus one link to
their list of specifications should be enough. We can keep the
others at the bottom, for now, but eventually I expect these to move
out to the implementation guide.

Co-authored-by: Jan Ainali <jan@publiccode.net>