Rather than hard-code port 9000, find a free one
The Ruby TCPServer.new() when given a port of 0 will
initialize to a valid available port.

This prints that and closes it, after which it is
available again for us to use.

There is, of course, a remote race condition where
another process grabs that same port between the
TCPServer.close() and the open() from Jekyll, but
that is also possible with a hard-coded port.