Replace "CHANGELOG" with "RELEASE_NOTES"
fixes #1049

The requirement text at the time of this writing is:

* The codebase SHOULD contain a log of changes from version to version,
  for example in the `CHANGELOG`.

However, this confuses the concepts of "release notes" and the "log of
changes".

Version-to-version changes are called "Release Notes" by The Open Source
Security Foundation Best Practices (formerly Core Infrastructure
Initiative Best Practices) as can be seen at:

https://www.bestpractices.dev/en/criteria

As we can see, the "Change Control" section includes:

> Release notes
>     * The project MUST provide, in each release, release notes that
>     are a human-readable summary of major changes in that release to
>     help users determine if they should upgrade and what the upgrade
>     impact will be.

Many software projects, notably the GNU projects, put their release
notes in a file named `NEWS`:

* https://www.gnu.org/prep/standards/html_node/NEWS-File.html

By contrast, a "ChangeLog" should be detailed enough and accurate enough
to provide the information commonly required for software forensics.  In
the modern era of widespread git usage this is something that is easily
provided by `git log`, yet if primarily distributed as a source tarball
release or if more narrative is desired, then a detailed "ChangeLog"
file may still be worth maintaining.  See also:

* https://www.gnu.org/prep/standards/html_node/Change-Logs.html
* https://www.gnu.org/prep/standards/html_node/Style-of-Change-Logs.html