Roadmap description adjustment
The contents of the roadmap were adjust slightly as part of 0.8.1
(see: 0b28d861c30639348d2d5f772182f404aacbe5cc),
however, the description at the top was not adjusted to match the
evolution to a comunity centered governance model.

This makes a first step towards reflecting the change in governance.

No roadmap items are changed as a part of this change.

Relates to: #1024